Move helper package to be an internal one
