+++ internal/helper/interface.go
